add filtering, sorting. add validation and error messages

+++ back/src/comment/comment.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from '../prisma.service';
 
 @Injectable()
@@ -11,22 +11,49 @@ export class CommentService {
     });
   }
 
-  async findOne(id: number) {
+  async findOne(id: string) {
     return this.prisma.comment.findUnique({
       where: { id },
     });
   }
 
-  async update(id: number, data: { text?: string; authorName?: string }) {
+  async update(id: string, data: { text?: string; authorName?: string }) {
+    const commentExists = await this.prisma.comment.findUnique({
+      where: {
+        id: id,
+      },
+    });
+
+    if (!commentExists) {
+      throw new NotFoundException(`Comment with id ${id} not found`);
+    }
+
     return this.prisma.comment.update({
       where: { id },
       data,
     });
   }
 
-  async remove(id: number) {
-    return this.prisma.comment.delete({
-      where: { id },
+  async remove(id: string): Promise<{ message: string; id?: string }> {
+    const commentExists = await this.prisma.comment.findUnique({
+      where: {
+        id: id,
+      },
     });
+
+    if (!commentExists) {
+      throw new NotFoundException(`Comment with id ${id} not found`);
+    }
+
+    await this.prisma.comment.delete({
+      where: {
+        id: id,
+      },
+    });
+
+    return {
+      message: 'Comment successfully deleted',
+      id: id,
+    };
   }
 }
